+++ fatfs/src/ops.rs
@@ -5,12 +5,12 @@ use glenda::error::Error;
 pub enum RootLocation {
     Cluster(u32),
     // sector, size_sectors
-    Sector(u64, u32),
+    Sector(usize, u32),
 }
 
 pub trait FatOps: Send + Sync {
     fn get_next_cluster(&self, reader: &BlockReader, cluster: u32) -> Result<u32, Error>;
-    fn cluster_to_sector(&self, cluster: u32) -> u64;
+    fn cluster_to_sector(&self, cluster: u32) -> usize;
     fn get_root_location(&self) -> RootLocation;
     fn bytes_per_sector(&self) -> u32;
     fn sectors_per_cluster(&self) -> u32;
